Fixing pylint
