+++ senscritique_export/row_utils/movies_utils.py
@@ -1,3 +1,4 @@
+"""Useful functions for movies items"""
 import logging
 from typing import Dict, List
