+++ senscritique_export/row_utils/music_utils.py
@@ -1,3 +1,4 @@
+"""Useful functions of music items"""
 import logging
 from typing import Dict, List
