+++ senscritique_export/row_utils/row_utils.py
@@ -1,3 +1,4 @@
+"""Useful functions for generic use"""
 import logging
 from typing import List, Optional
 
@@ -21,8 +22,8 @@ def parse_baseline(row: element.Tag) -> Optional[List[str]]:
     """
     try:
         return row.find("p", {"class": "elco-baseline"}).text.replace("\n", "").replace("\t", "").split(".")
-    except Exception as e:
-        logger.debug("Function parse_baseline for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function parse_baseline for row %s : %s", row, exception)
         return None
 
 
@@ -41,8 +42,8 @@ def get_baseline_0(row: element.Tag) -> Optional[str]:
     """
     try:
         return parse_baseline(row)[0].strip()
-    except Exception as e:
-        logger.debug("Function get_baseline_0 for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_baseline_0 for row %s : %s", row, exception)
         return None
 
 
@@ -61,8 +62,8 @@ def get_baseline_1(row: element.Tag) -> Optional[str]:
     """
     try:
         return parse_baseline(row)[1].strip()
-    except Exception as e:
-        logger.debug("Function get_baseline_1 for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_baseline_1 for row %s : %s", row, exception)
         return None
 
 
@@ -81,8 +82,8 @@ def get_baseline_2(row: element.Tag) -> Optional[str]:
     """
     try:
         return parse_baseline(row)[2].strip()
-    except Exception as e:
-        logger.debug("Function get_baseline_2 for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_baseline_2 for row %s : %s", row, exception)
         return None
 
 
@@ -102,12 +103,11 @@ def get_rank(row: element.Tag) -> Optional[str]:
     try:
         if row.find("span", {"class": "elpo-rank-item"}):
             return row.find("span", {"class": "elpo-rank-item"}).text
-        elif row.find("span", {"class": "elto-rank-item"}):
+        if row.find("span", {"class": "elto-rank-item"}):
             return row.find("span", {"class": "elto-rank-item"}).text
-        else:
-            return None
-    except Exception as e:
-        logger.debug("Function get_rank for row %s : %s", row, e)
+        return None
+    except Exception as exception:
+        logger.debug("Function get_rank for row %s : %s", row, exception)
         return None
 
 
@@ -126,8 +126,8 @@ def get_title(row: element.Tag) -> Optional[str]:
     """
     try:
         return row.find("a", {"class": "elco-anchor"}).text.strip()
-    except Exception as e:
-        logger.debug("Function get_title for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_title for row %s : %s", row, exception)
         return None
 
 
@@ -146,8 +146,8 @@ def get_url(row: element.Tag) -> Optional[str]:
     """
     try:
         return "https://old.senscritique.com" + row.find("a", {"class": "elco-anchor"})["href"]
-    except Exception as e:
-        logger.debug("Function get_url for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_url for row %s : %s", row, exception)
         return None
 
 
@@ -169,8 +169,8 @@ def get_original_title(row: element.Tag) -> Optional[str]:
             original_title = row.find("p", {"class": "elco-original-title"}).text.strip()
         else:
             original_title = None
-    except Exception as e:
-        logger.debug("Function get_original_title for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_original_title for row %s : %s", row, exception)
         return None
     return original_title
 
@@ -190,8 +190,8 @@ def get_year(row: element.Tag) -> Optional[str]:
     """
     try:
         return row.find("span", {"class": "elco-date"}).text.replace("(", "").replace(")", "")
-    except Exception as e:
-        logger.debug("Function get_year for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_year for row %s : %s", row, exception)
         return None
 
 
@@ -212,13 +212,13 @@ def get_picture_url(row: element.Tag) -> Optional[str]:
         if row.select("img"):
             try:
                 picture_url = row.find("img")["src"]
-            except Exception as e:
-                logger.debug(e)
+            except Exception as exception:
+                logger.debug(exception)
                 picture_url = row.find("img")["data-original"]
         else:
             picture_url = None
-    except Exception as e:
-        logger.debug("Function get_picture_url for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_picture_url for row %s : %s", row, exception)
         picture_url = None
     return picture_url
 
@@ -241,8 +241,8 @@ def get_genre(row: element.Tag) -> Optional[str]:
             genre = parse_baseline(row)[2].strip()
         else:
             genre = parse_baseline(row)[3].strip()
-    except Exception as e:
-        logger.debug("Function get_genre for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_genre for row %s : %s", row, exception)
         genre = None
     return genre
 
@@ -265,8 +265,8 @@ def get_producer(row: element.Tag) -> Optional[str]:
             producer = ", ".join([x.text.strip() for x in row.find_all("span", {"class": "elco-baseline-a"})])
         else:
             producer = ", ".join([x.text.strip() for x in row.find_all("a", {"class": "elco-baseline-a"})])
-    except Exception as e:
-        logger.debug("Function get_producer for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_producer for row %s : %s", row, exception)
         producer = None
     return producer
 
@@ -286,8 +286,8 @@ def get_description(row: element.Tag) -> Optional[str]:
     """
     try:
         description = row.find("p", {"class": "elco-description"}).text.strip()
-    except Exception as e:
-        logger.debug("Function get_description for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_description for row %s : %s", row, exception)
         description = None
     return description
 
@@ -307,8 +307,8 @@ def get_average_rating(row: element.Tag) -> Optional[str]:
     """
     try:
         average_rating = row.find("a", {"class": "erra-global"}).text.strip()
-    except Exception as e:
-        logger.debug("Function get_average_rating for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_average_rating for row %s : %s", row, exception)
         average_rating = None
     return average_rating
 
@@ -328,8 +328,8 @@ def get_number_of_ratings(row: element.Tag) -> Optional[str]:
     """
     try:
         number_of_ratings = row.find("a", {"class": "erra-global"})["title"].split()[-2]
-    except Exception as e:
-        logger.debug("Function get_number_of_ratings for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_number_of_ratings for row %s : %s", row, exception)
         number_of_ratings = None
     return number_of_ratings
 
@@ -351,8 +351,8 @@ def get_number_of_seasons(row: element.Tag) -> Optional[str]:
         number_of_seasons = parse_baseline(row)[2].strip()
         if not any(i.isdigit() for i in number_of_seasons):
             return None
-    except Exception as e:
-        logger.debug("Function get_number_of_seasons for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_number_of_seasons for row %s : %s", row, exception)
         number_of_seasons = None
     return number_of_seasons
 
@@ -372,8 +372,8 @@ def get_platforms(row: element.Tag) -> Optional[str]:
     """
     try:
         platforms = row.find("span", {"class": "elco-gamesystem"}).text.strip()
-    except Exception as e:
-        logger.debug("Function get_platforms for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_platforms for row %s : %s", row, exception)
         platforms = None
     return platforms
 
@@ -393,7 +393,7 @@ def get_topchart_platforms(row: element.Tag) -> Optional[str]:
     """
     try:
         platforms = row.find_all("p", {"class": "elco-baseline"})[1].text.split("sur")[-1].strip()
-    except Exception as e:
-        logger.debug("Function get_topchart_platforms for row %s : %s", row, e)
+    except Exception as exception:
+        logger.debug("Function get_topchart_platforms for row %s : %s", row, exception)
         platforms = None
     return platforms
