+++ senscritique_export/row_utils/series_utils.py
@@ -1,3 +1,4 @@
+"""Useful functions for series items"""
 import logging
 from typing import Dict, List
