+++ tox.ini
@@ -94,8 +94,8 @@ commands =
 deps =
     pylint~=2.14.4
 commands =
-    python -m pylint senscritique_export --rcfile .pylintrc
-    python -m pylint tests --rcfile .pylintrc --disable C0114,C0115,C0116,R0201,R0903,W0106,W0621
+    python -m pylint senscritique_export --rcfile .pylintrc --disable R0801,W0201,W0703
+    python -m pylint tests --rcfile .pylintrc --disable C0114,C0115,C0116,R0903,W0106,W0621
 
 [testenv:spelling]
 deps =
